copy to server: impl `move_in_same_fs_to` mode

+++ src/client.rs
@@ -31,8 +31,19 @@ pub(crate) struct Config {
 #[derive(Deserialize, Debug)]
 #[serde(untagged)]
 enum CopyToServer {
+    Move { move_in_same_fs_to: PathBuf },
     Copy { destination: PathBuf },
     Command(Vec<String>),
+}
+
+impl CopyToServer {
+    fn requires_cleanup(&self) -> bool {
+        match self {
+            CopyToServer::Move { .. } => false,
+            CopyToServer::Copy { .. } => true,
+            CopyToServer::Command(_) => true,
+        }
+    }
 }
 
 #[derive(Deserialize, Debug)]
@@ -101,12 +112,14 @@ async fn listen_to_server(
         match msg {
             Receipt::Received(spec) => {
                 info!("server confirmed reception of {spec:?}");
-                let path = conf.watched_path(&spec);
-                if let Err(err) = fs::remove_file(&path).await {
-                    warn!("error when removing {path:?}: {err}");
-                } else {
-                    db.lock().await.remove(&spec.relative_path());
+                if conf.copy_to_server.requires_cleanup() {
+                    let path = conf.watched_path(&spec);
+                    if let Err(err) = fs::remove_file(&path).await {
+                        warn!("error when removing {path:?}: {err}");
+                        continue;
+                    }
                 }
+                db.lock().await.remove(&spec.relative_path());
             }
             Receipt::DifferentHash { spec, .. } => {
                 info!("server does not have expected hash for {spec:?}, resending");
@@ -156,6 +169,12 @@ async fn send_file_to_server(
 ) {
     let from = conf.watched_path(&spec);
     let outcome = match &conf.copy_to_server {
+        CopyToServer::Move { move_in_same_fs_to } => {
+            info!("move {spec:?} to server via `fs::rename`");
+            let mut destination = move_in_same_fs_to.clone();
+            destination.push(spec.server_filename());
+            fs::rename(from, destination).await.map(|_| ()).into()
+        }
         CopyToServer::Copy { destination } => {
             info!("copying {spec:?} to server via `fs::copy`");
             let mut destination = destination.clone();
